+++ CHANGELOG.md
@@ -4,7 +4,6 @@ All notable changes to this project will be documented in this file. See [standa
 
 ### 0.1.1 (2022-04-07)
 
-
 ### Features
 
-* **global:** initialize project ([efe4e7e](https://github.com/kioshiokamoto/crowdfunding-front/commit/efe4e7ead4a53acb74ccde2e868fb4bb5dd717fb))
+- **global:** initialize project ([efe4e7e](https://github.com/kioshiokamoto/crowdfunding-front/commit/efe4e7ead4a53acb74ccde2e868fb4bb5dd717fb))